refactor: moved recruit item component folder from recruit page folder to UI folder
